requirements.txt
opencv-python
Pillow
numpy
bleak
